% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenmat_class.R
\name{as.double.Tenmat}
\alias{as.double.Tenmat}
\title{Convert Tenmat to standard R Matrix using generic type conversion}
\usage{
\method{as.double}{Tenmat}(x, ...)
}
\arguments{
\item{x}{A Tenmat object}

\item{...}{Additional arguments (ignored)}
}
\value{
A double matrix
}
\description{
Convert Tenmat to standard R Matrix using generic type conversion
}
