% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenmat_class.R
\name{as.matrix.Tenmat}
\alias{as.matrix.Tenmat}
\title{Convert Tenmat to standard R Matrix}
\usage{
\method{as.matrix}{Tenmat}(x, ...)
}
\arguments{
\item{x}{A Tenmat object}

\item{...}{Additional arguments (ignored)}
}
\value{
A matrix
}
\description{
Convert Tenmat to standard R Matrix
}
