% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenmat_class.R
\name{as.tenmat}
\alias{as.tenmat}
\title{Convert object to Tenmat}
\usage{
as.tenmat(x)
}
\arguments{
\item{x}{Object to convert}
}
\value{
A Tenmat object
}
\description{
Convert object to Tenmat
}
\examples{
t <- tensor(array(1:24, dim = c(3, 4, 2)))
as.tenmat(t)

}
