% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ktensor_class.R
\name{as.tensor.KTensor}
\alias{as.tensor.KTensor}
\title{S3 function to convert KTensor to full Tensor}
\usage{
\method{as.tensor}{KTensor}(x, ...)
}
\arguments{
\item{x}{A KTensor object}

\item{...}{Additional arguments}
}
\value{
A dense Tensor object
}
\description{
S3 function to convert KTensor to full Tensor
}
