% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_class.R
\name{as.tensor}
\alias{as.tensor}
\title{Convert object to Tensor}
\usage{
as.tensor(x, ...)
}
\arguments{
\item{x}{Object to convert}

\item{...}{Additional arguments}
}
\value{
A Tensor object
}
\description{
Convert object to Tensor
}
