% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenmat_class.R
\name{as.tensor.Tenmat}
\alias{as.tensor.Tenmat}
\title{Convert Tenmat to Tensor}
\usage{
\method{as.tensor}{Tenmat}(x, ...)
}
\arguments{
\item{x}{A Tenmat object}

\item{...}{Additional arguments (ignored)}
}
\value{
A Tensor object
}
\description{
Convert Tenmat to Tensor
}
\examples{
t <- tensor(array(1:24, dim = c(3, 4, 2)))
tm <- tenmat(t, rdims = 1)
t2 <- as.tensor(tm)

}
