% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttensor_class.R
\name{as.tensor.TTensor}
\alias{as.tensor.TTensor}
\title{S3 function to convert TTensor to full Tensor}
\usage{
\method{as.tensor}{TTensor}(x, ...)
}
\arguments{
\item{x}{A TTensor object}

\item{...}{Additional arguments}
}
\value{
A dense Tensor object
}
\description{
S3 function to convert TTensor to full Tensor
}
