% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenmat_class.R
\name{as.vector.Tenmat}
\alias{as.vector.Tenmat}
\title{Convert Tenmat to standard R vector}
\usage{
\method{as.vector}{Tenmat}(x, mode = "any")
}
\arguments{
\item{x}{A Tenmat object}

\item{mode}{The type of vector to be returned (e.g., "numeric", "character", "logical").}
}
\value{
A vector
}
\description{
Convert Tenmat to standard R vector
}
