% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenmat_class.R
\name{double.Tenmat}
\alias{double.Tenmat}
\title{Convert Tenmat to standard R double array (alias for matrix)}
\usage{
double.Tenmat(x)
}
\arguments{
\item{x}{A Tenmat object}
}
\value{
A double matrix
}
\description{
Convert Tenmat to standard R double array (alias for matrix)
}
