% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_dense_methods.R
\name{double.Tensor}
\alias{double.Tensor}
\title{MATLAB-Style Double Conversion}
\usage{
double.Tensor(x)
}
\arguments{
\item{x}{A Tensor object.}
}
\value{
A base R array.
}
\description{
Returns the dense array data stored by a tensor.
}
