% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_operations.R
\name{fnorm}
\alias{fnorm}
\title{Frobenius Norm}
\usage{
fnorm(x, ...)
}
\arguments{
\item{x}{A tensor or matrix.}

\item{...}{Additional arguments.}
}
\value{
A scalar value representing the Frobenius norm.
}
\description{
Calculates the Frobenius norm of a tensor.
}
