% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_dense_methods.R
\name{full}
\alias{full}
\title{Dense Array Representation}
\usage{
full(x, ...)
}
\arguments{
\item{x}{A Tensor object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A base R array.
}
\description{
Returns the dense base-array representation of a tensor.
}
