% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenmat_class.R
\name{\%*\%}
\alias{\%*\%}
\alias{\%*\%.Tenmat}
\title{S3 Matrix Multiplication Generic}
\usage{
x \%*\% y

\method{\%*\%}{Tenmat}(x, y)
}
\arguments{
\item{x}{First argument}

\item{y}{Second argument}
}
\value{
Result of matrix multiplication
}
\description{
S3 Matrix Multiplication Generic
}
