% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_operations.R
\name{hadamard}
\alias{hadamard}
\title{Hadamard Product}
\usage{
hadamard(x, ...)
}
\arguments{
\item{x}{A matrix or list of matrices.}

\item{...}{Additional arguments.}

\item{y}{A matrix (if x is a matrix).}
}
\value{
A matrix representing the Hadamard product.
}
\description{
Element-wise multiplication of matrices or list of matrices.
}
