% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_class.R
\name{head.Tensor}
\alias{head.Tensor}
\title{S3 head method for Tensor}
\usage{
\method{head}{Tensor}(x, ...)
}
\arguments{
\item{x}{A Tensor object}

\item{...}{Additional arguments}
}
\description{
S3 head method for Tensor
}
