% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_constructors.R
\name{import_data}
\alias{import_data}
\title{Import Tensor Data from a Text File}
\usage{
import_data(fname)
}
\arguments{
\item{fname}{Path of the file to read.}
}
\value{
A \code{Tensor}, matrix, \code{KTensor}, or \code{Sptensor} depending on the file
header.
}
\description{
Reads a file written by \code{\link[=export_data]{export_data()}} (or MATLAB Tensor Toolbox
\code{exportdata}) and reconstructs the corresponding object.
}
\seealso{
\code{\link[=export_data]{export_data()}}
}
