% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_dense_methods.R
\name{isequal}
\alias{isequal}
\title{Equality Test for Tensors}
\usage{
isequal(x, y, ...)
}
\arguments{
\item{x}{First object.}

\item{y}{Second object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Logical scalar.
}
\description{
Equality Test for Tensors
}
