% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_dense_methods.R
\name{isscalar}
\alias{isscalar}
\title{Scalar Tensor Predicate}
\usage{
isscalar(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Logical scalar.
}
\description{
Scalar Tensor Predicate
}
