% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_dense_methods.R
\name{issymmetric}
\alias{issymmetric}
\title{Check Tensor Symmetry}
\usage{
issymmetric(x, grps = NULL, ...)
}
\arguments{
\item{x}{A Tensor object.}

\item{grps}{A vector of modes or list of mode vectors.}
}
\value{
Logical scalar.
}
\description{
Checks whether a tensor is symmetric with respect to one or more groups of
modes.
}
