% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_class.R
\name{Math.Tensor}
\alias{Math.Tensor}
\title{S3 Math group generic for Tensor}
\usage{
\method{Math}{Tensor}(x, ...)
}
\arguments{
\item{x}{A Tensor object}

\item{...}{Additional arguments}
}
\description{
S3 Math group generic for Tensor
}
