% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenmat_class.R
\name{mtimes}
\alias{mtimes}
\title{Matrix Multiplication Alias}
\usage{
mtimes(x, y)
}
\arguments{
\item{x}{First argument}

\item{y}{Second argument}
}
\value{
Result of matrix multiplication
}
\description{
Matrix Multiplication Alias
}
