% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_dense_methods.R
\name{mttkrp}
\alias{mttkrp}
\title{Matricized Tensor Times Khatri-Rao Product}
\usage{
mttkrp(x, U, mode, ...)
}
\arguments{
\item{x}{A Tensor object.}

\item{U}{A list of factor matrices or a \code{KTensor}.}

\item{mode}{Mode to skip.}
}
\value{
A matrix of size \verb{dim(x)[mode] x R}.
}
\description{
Computes the mode-n unfolding multiplied by the Khatri-Rao product of all
factor matrices except the skipped mode.
}
