% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_dense_methods.R
\name{mttkrps}
\alias{mttkrps}
\title{Sequence of MTTKRP Calculations}
\usage{
mttkrps(x, U, ...)
}
\arguments{
\item{x}{A Tensor object.}

\item{U}{A list of factor matrices or a \code{KTensor}.}
}
\value{
A list of matrices, one per mode.
}
\description{
Computes \code{mttkrp(x, U, k)} for all tensor modes.
}
