% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ktensor_methods.R
\name{ncomponents}
\alias{ncomponents}
\title{Number of Components of a Kruskal Tensor}
\usage{
ncomponents(x, ...)
}
\arguments{
\item{x}{A \code{KTensor}.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Integer number of rank-one components.
}
\description{
Number of Components of a Kruskal Tensor
}
