% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_dense_methods.R
\name{nnz}
\alias{nnz}
\title{Number of Nonzeros}
\usage{
nnz(x, ...)
}
\arguments{
\item{x}{A tensor-like object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Integer count of nonzero entries.
}
\description{
Number of Nonzeros
}
