% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_class.R
\name{ones}
\alias{ones}
\title{Create a tensor of ones}
\usage{
ones(dims)
}
\arguments{
\item{dims}{Dimensions of the tensor}
}
\value{
A new Tensor object filled with ones
}
\description{
Create a tensor of ones
}
