% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ktensor_class.R
\name{print.KTensor}
\alias{print.KTensor}
\title{S3 print method for KTensor}
\usage{
\method{print}{KTensor}(x, ...)
}
\arguments{
\item{x}{A KTensor object}

\item{...}{Additional arguments}
}
\description{
S3 print method for KTensor
}
