% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_dense_methods.R
\name{reshape}
\alias{reshape}
\title{Reshape Tensor}
\usage{
reshape(x, new_dims, ...)
}
\arguments{
\item{x}{A tensor-like object.}

\item{new_dims}{New dimensions.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A reshaped object.
}
\description{
Reshape Tensor
}
