% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptensor_class.R
\name{Sptensor}
\alias{Sptensor}
\alias{sptensor}
\title{R6 Class for Sparse Tensors (Sptensor)}
\usage{
sptensor(subs = NULL, vals = NULL, dims = NULL)
}
\arguments{
\item{subs}{Integer matrix of subscripts (\verb{nnz x ndims}), or a dense
Tensor/array to convert.}

\item{vals}{Numeric vector of values, one per subscript row.}

\item{dims}{Integer vector of tensor dimensions.}
}
\value{
An \code{Sptensor} object.
}
\description{
Stores a tensor as a list of subscripts and nonzero values, mirroring the
MATLAB Tensor Toolbox \code{sptensor}. Duplicate subscripts are summed and
explicit zeros dropped on construction.
}
\examples{
S <- sptensor(rbind(c(1, 1, 1), c(2, 3, 4)), c(5, 7), c(2, 3, 4))
}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{subs}}{Integer matrix of subscripts, one row per nonzero}

    \item{\code{vals}}{Numeric vector of nonzero values}

    \item{\code{dims}}{Integer vector of dimensions}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-Sptensor-initialize}{\code{Sptensor$new()}}
    \item \href{#method-Sptensor-dim}{\code{Sptensor$dim()}}
    \item \href{#method-Sptensor-ndims}{\code{Sptensor$ndims()}}
    \item \href{#method-Sptensor-print}{\code{Sptensor$print()}}
    \item \href{#method-Sptensor-full}{\code{Sptensor$full()}}
    \item \href{#method-Sptensor-clone}{\code{Sptensor$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sptensor-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-Sptensor-initialize}{}}}
\subsection{\code{Sptensor$new()}}{
  Initialize a new Sptensor
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Sptensor$new(subs = NULL, vals = NULL, dims = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{subs}}{Integer matrix of subscripts (\verb{nnz x ndims}).}
      \item{\code{vals}}{Numeric vector of values, one per subscript row (a scalar
is recycled).}
      \item{\code{dims}}{Integer vector of tensor dimensions.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A new Sptensor object
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sptensor-dim"></a>}}
\if{latex}{\out{\hypertarget{method-Sptensor-dim}{}}}
\subsection{\code{Sptensor$dim()}}{
  Get dimensions of the tensor
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Sptensor$dim()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Integer vector of dimensions
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sptensor-ndims"></a>}}
\if{latex}{\out{\hypertarget{method-Sptensor-ndims}{}}}
\subsection{\code{Sptensor$ndims()}}{
  Get number of dimensions
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Sptensor$ndims()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Integer tensor order
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sptensor-print"></a>}}
\if{latex}{\out{\hypertarget{method-Sptensor-print}{}}}
\subsection{\code{Sptensor$print()}}{
  Print the Sptensor object
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Sptensor$print(...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{...}}{Additional arguments}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Invisible self
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sptensor-full"></a>}}
\if{latex}{\out{\hypertarget{method-Sptensor-full}{}}}
\subsection{\code{Sptensor$full()}}{
  Convert to a dense Tensor
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Sptensor$full()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A dense \code{Tensor}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sptensor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Sptensor-clone}{}}}
\subsection{\code{Sptensor$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Sptensor$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
