% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_class.R
\name{Summary.Tensor}
\alias{Summary.Tensor}
\title{S3 Summary group generic for Tensor}
\usage{
\method{Summary}{Tensor}(..., na.rm = FALSE)
}
\arguments{
\item{...}{Tensor objects or numeric values}

\item{na.rm}{Logical indicating whether missing values should be removed}
}
\description{
S3 Summary group generic for Tensor
}
