% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_dense_methods.R
\name{symmetrize}
\alias{symmetrize}
\title{Symmetrize Tensor}
\usage{
symmetrize(x, grps = NULL, ...)
}
\arguments{
\item{x}{A Tensor object.}

\item{grps}{A vector of modes or list of mode vectors.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A symmetrized Tensor object.
}
\description{
Symmetrizes a tensor with respect to one or more mode groups.
}
