% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_operations.R
\name{t_scale}
\alias{t_scale}
\title{Scale Tensor}
\usage{
t_scale(x, ...)
}
\arguments{
\item{x}{A Tensor.}

\item{...}{Additional arguments.}

\item{s}{A vector or Tensor containing scaling factors.}

\item{dims}{Dimensions to scale.}
}
\value{
A scaled Tensor.
}
\description{
Scales a tensor along specified dimensions. Equivalent to broadcasting
a scaling vector or tensor for multiplication.
}
