% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_class.R
\name{tail.Tensor}
\alias{tail.Tensor}
\title{S3 tail method for Tensor}
\usage{
\method{tail}{Tensor}(x, ...)
}
\arguments{
\item{x}{A Tensor object}

\item{...}{Additional arguments}
}
\description{
S3 tail method for Tensor
}
