% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_dense_methods.R
\name{tenfun}
\alias{tenfun}
\title{Apply Elementwise Function to Tensor Arguments}
\usage{
tenfun(fun, ...)
}
\arguments{
\item{fun}{A function or function-like object.}

\item{...}{Tensor-like objects or scalars.}
}
\value{
A Tensor object.
}
\description{
Apply Elementwise Function to Tensor Arguments
}
