% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenmat_class.R
\name{tenmat}
\alias{tenmat}
\title{Create a tenmat (tensor as matrix) object}
\usage{
tenmat(T, rdims = NULL, cdims = NULL, tsize = NULL)
}
\arguments{
\item{T}{A Tensor object, matrix, or Tenmat object}

\item{rdims}{Dimensions to map to rows}

\item{cdims}{Dimensions to map to columns (optional). Also accepts 't', 'fc', 'bc'.}

\item{tsize}{Tensor size (used internally)}
}
\value{
A new Tenmat object
}
\description{
Create a tenmat (tensor as matrix) object
}
