% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_class.R
\docType{class}
\name{tensor-class}
\alias{tensor-class}
\title{S4 `tensor` class}
\description{
S4 `tensor` class
}
\section{Slots}{

\describe{
\item{\code{data}}{An R  array.}

\item{\code{dims}}{The dimension of the tensor.}
}}

