% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_dense_methods.R
\name{transpose}
\alias{transpose}
\title{Transpose Tensor}
\usage{
transpose(x, ...)
}
\arguments{
\item{x}{A Tensor object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
This function always errors for tensors.
}
\description{
Tensor transpose is not defined; use \code{\link[=permute]{permute()}} instead.
}
