% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_dense_methods.R
\name{ttv}
\alias{ttv}
\title{Tensor Times Vector}
\usage{
ttv(tensor, vector, mode = NULL)
}
\arguments{
\item{tensor}{A Tensor-like object.}

\item{vector}{A numeric vector or list of numeric vectors.}

\item{mode}{Integer mode or modes to contract.}
}
\value{
A Tensor object.
}
\description{
Convenience wrapper around \code{\link[=ttm]{ttm()}} for vector contractions.
}
