% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_dense_methods.R
\name{unfold}
\alias{unfold}
\title{Unfold Tensor}
\usage{
unfold(x, rdims = NULL, cdims = NULL, ...)
}
\arguments{
\item{x}{A tensor-like object.}

\item{rdims}{Row modes.}

\item{cdims}{Column modes. If omitted, all remaining modes are used.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A matrix.
}
\description{
Unfold Tensor
}
