% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_dense_methods.R
\name{vec}
\alias{vec}
\title{Vectorize Tensor}
\usage{
vec(x, ...)
}
\arguments{
\item{x}{A tensor-like object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A vector.
}
\description{
Vectorize Tensor
}
