% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_class.R
\name{zeros}
\alias{zeros}
\title{Create a tensor of zeros}
\usage{
zeros(dims)
}
\arguments{
\item{dims}{Dimensions of the tensor}
}
\value{
A new Tensor object filled with zeros
}
\description{
Create a tensor of zeros
}
