% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_operations.R
\name{collapse}
\alias{collapse}
\alias{collapse.Tensor}
\title{Collapse Tensor}
\usage{
collapse(x, ...)

\method{collapse}{Tensor}(x, dims, fun = sum, ...)
}
\arguments{
\item{x}{A Tensor object.}

\item{...}{Additional arguments passed to \code{fun}.}

\item{dims}{Integer vector specifying the dimensions to collapse along.
Negative dimensions exclude those dimensions.}

\item{fun}{The function to apply (default is sum).}
}
\value{
A collapsed Tensor.
}
\description{
Collapses a tensor over specified dimensions using an accumulation function
(e.g., sum, mean, max).
}
\examples{
t <- tensor(array(1:24, dim = c(3, 4, 2)))
collapse(t, dims = 1) # Collapse the first dimension
collapse(t, dims = c(1, 2), fun = mean) # Mean over first two dimensions

}
