% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_dense_methods.R
\name{contract}
\alias{contract}
\title{Contract Tensor Dimensions}
\usage{
contract(x, i, j, ...)
}
\arguments{
\item{x}{A Tensor object.}

\item{i}{First mode.}

\item{j}{Second mode.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A Tensor object, potentially scalar.
}
\description{
Contracts a tensor along two equally sized dimensions.
}
