% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_dense_methods.R
\name{fibers}
\alias{fibers}
\title{Extract Tensor Fibers}
\usage{
fibers(x, mode, midx, ...)
}
\arguments{
\item{x}{A Tensor object.}

\item{mode}{Mode of the fibers to extract.}

\item{midx}{Matrix with one row per requested fiber and \code{ndims(x) - 1}
columns containing the fixed indices for the other modes.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A matrix of size \verb{dim(x)[mode] x nrow(midx)}.
}
\description{
Extracts mode-k fibers specified by a matrix of sample indices.
}
