% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_operations.R
\name{hadamard}
\alias{hadamard}
\title{Hadamard Product}
\usage{
hadamard(x, ...)
}
\arguments{
\item{x}{A matrix or list of matrices.}

\item{...}{Additional arguments.}

\item{y}{A matrix (if x is a matrix).}
}
\value{
A matrix representing the Hadamard product.
}
\description{
Element-wise multiplication of matrices or list of matrices.
}
\examples{
A <- matrix(1:4, nrow = 2)
B <- matrix(5:8, nrow = 2)
hadamard(A, B)

# With a list of matrices
C <- matrix(9:12, nrow = 2)
hadamard(list(A, B, C))

}
