% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_operations.R
\name{khatri_rao}
\alias{khatri_rao}
\title{Khatri-Rao Product}
\usage{
khatri_rao(x, ...)
}
\arguments{
\item{x}{A matrix or list of matrices.}

\item{...}{Additional arguments.}

\item{y}{A matrix (if x is a matrix).}

\item{reverse}{Logical. If TRUE, computes y \eqn{\odot} x instead of x \eqn{\odot} y.}
}
\value{
A matrix representing the Khatri-Rao product.
}
\description{
Computes the Khatri-Rao product (column-wise Kronecker product) of two matrices,
or a list of matrices.
}
