% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_operations.R
\name{khatri_rao}
\alias{khatri_rao}
\alias{khatri_rao.matrix}
\title{Khatri-Rao Product}
\usage{
khatri_rao(x, ...)

\method{khatri_rao}{matrix}(x, y, reverse = FALSE, ...)
}
\arguments{
\item{x}{A matrix or list of matrices.}

\item{...}{Additional arguments.}

\item{y}{A matrix (if x is a matrix).}

\item{reverse}{Logical. If TRUE, computes y \eqn{\odot} x instead of x \eqn{\odot} y.}
}
\value{
A matrix representing the Khatri-Rao product.
}
\description{
Computes the Khatri-Rao product (column-wise Kronecker product) of two matrices,
or a list of matrices.
}
\examples{
A <- matrix(1:4, nrow = 2)
B <- matrix(5:8, nrow = 2)
khatri_rao(A, B)

# With a list of matrices
C <- matrix(9:12, nrow = 2)
khatri_rao(list(A, B, C))

}
