% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_operations.R
\name{kronecker}
\alias{kronecker}
\title{Kronecker Product}
\usage{
kronecker(X, Y = NULL, FUN = "*", make.dimnames = FALSE, ...)
}
\arguments{
\item{X}{A matrix, Tensor, or list of matrices.}

\item{Y}{A matrix or Tensor (optional if X is a list).}

\item{FUN}{The function to use (default "*").}

\item{make.dimnames}{Logical.}

\item{...}{Additional arguments.}
}
\value{
The Kronecker product.
}
\description{
Extends base::kronecker to support list of matrices.
}
\examples{
A <- matrix(1:4, nrow = 2)
B <- matrix(5:8, nrow = 2)
kronecker(A, B)

# With a list of matrices
C <- matrix(9:12, nrow = 2)
kronecker(list(A, B, C))

}
