% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ktensor_class.R
\name{ktensor}
\alias{ktensor}
\title{Create a Kruskal tensor (KTensor)}
\usage{
ktensor(lambda = NULL, U = NULL)
}
\arguments{
\item{lambda}{A numeric vector of weights, or list of matrices U if lambda is omitted.}

\item{U}{A list of factor matrices. Let NULL if passing list as \code{lambda}.}
}
\value{
A KTensor object
}
\description{
Create a Kruskal tensor (KTensor)
}
