% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ktensor_class.R
\name{KTensor}
\alias{KTensor}
\title{R6 Class for Kruskal Tensors (KTensor)}
\description{
A class representing Kruskal tensors (KTensor), which are the sum of outer products
of vectors. This matches the behavior of MATLAB's Tensor Toolbox ktensor.

A Kruskal tensor is represented by a set of factor matrices and a vector of weights.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{lambda}}{Vector of weights}

\item{\code{U}}{List of factor matrices}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-KTensor-new}{\code{KTensor$new()}}
\item \href{#method-KTensor-dim}{\code{KTensor$dim()}}
\item \href{#method-KTensor-ndims}{\code{KTensor$ndims()}}
\item \href{#method-KTensor-print}{\code{KTensor$print()}}
\item \href{#method-KTensor-full}{\code{KTensor$full()}}
\item \href{#method-KTensor-clone}{\code{KTensor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KTensor-new"></a>}}
\if{latex}{\out{\hypertarget{method-KTensor-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new KTensor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KTensor$new(lambda = NULL, U = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lambda}}{A numeric vector of weights. If NULL, defaults to a vector of 1s.}

\item{\code{U}}{A list of factor matrices}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new KTensor object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KTensor-dim"></a>}}
\if{latex}{\out{\hypertarget{method-KTensor-dim}{}}}
\subsection{Method \code{dim()}}{
Get dimensions of the tensor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KTensor$dim()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Integer vector of dimensions
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KTensor-ndims"></a>}}
\if{latex}{\out{\hypertarget{method-KTensor-ndims}{}}}
\subsection{Method \code{ndims()}}{
Get number of dimensions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KTensor$ndims()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Integer number of dimensions
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KTensor-print"></a>}}
\if{latex}{\out{\hypertarget{method-KTensor-print}{}}}
\subsection{Method \code{print()}}{
Print the KTensor object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KTensor$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KTensor-full"></a>}}
\if{latex}{\out{\hypertarget{method-KTensor-full}{}}}
\subsection{Method \code{full()}}{
Convert Kruskal tensor to dense Tensor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KTensor$full()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new Tensor object representing the full dense tensor
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KTensor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-KTensor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KTensor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
