% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_dense_methods.R
\name{mask}
\alias{mask}
\title{Mask Tensor Values}
\usage{
mask(x, w, ...)
}
\arguments{
\item{x}{A Tensor object.}

\item{w}{A Tensor-like mask.}
}
\value{
A vector of extracted values.
}
\description{
Extracts values from a tensor at positions corresponding to nonzero entries
of a mask tensor.
}
