% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_dense_methods.R
\name{permute}
\alias{permute}
\title{Permute Tensor Dimensions}
\usage{
permute(x, order, ...)
}
\arguments{
\item{x}{A tensor-like object.}

\item{order}{Permutation of dimensions.}
}
\value{
A permuted Tensor object.
}
\description{
Permute Tensor Dimensions
}
