% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_dense_methods.R
\name{squeeze}
\alias{squeeze}
\title{Squeeze Tensor}
\usage{
squeeze(x, ...)
}
\arguments{
\item{x}{A tensor-like object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Object with singleton dimensions removed.
}
\description{
Squeeze Tensor
}
