% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_operations.R
\name{t_scale}
\alias{t_scale}
\alias{t_scale.Tensor}
\title{Scale Tensor}
\usage{
t_scale(x, ...)

\method{t_scale}{Tensor}(x, s, dims, ...)
}
\arguments{
\item{x}{A Tensor.}

\item{...}{Additional arguments.}

\item{s}{A vector or Tensor containing scaling factors.}

\item{dims}{Dimensions to scale.}
}
\value{
A scaled Tensor.
}
\description{
Scales a tensor along specified dimensions. Equivalent to broadcasting
a scaling vector or tensor for multiplication.
}
\examples{
t <- tensor(array(1:24, dim = c(3, 4, 2)))
s <- c(10, 100, 1000)
t_scale(t, s, dims = 1) # Scale along the first dimension

}
