% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenmat_class.R
\name{tenmat}
\alias{tenmat}
\title{Create a tenmat (tensor as matrix) object}
\usage{
tenmat(T, rdims = NULL, cdims = NULL, tsize = NULL)
}
\arguments{
\item{T}{A Tensor object, matrix, or Tenmat object}

\item{rdims}{Dimensions to map to rows}

\item{cdims}{Dimensions to map to columns (optional). Also accepts 't', 'fc', 'bc'.}

\item{tsize}{Tensor size (used internally)}
}
\value{
A new Tenmat object
}
\description{
Create a tenmat (tensor as matrix) object
}
\examples{
t <- tensor(array(1:24, dim = c(3, 4, 2)))
# Unfold mode 1 to rows, modes 2 and 3 to columns
m <- tenmat(t, rdims = 1)
print(m)

# Unfold mode 2 to rows, and mode 3 and 1 to columns (forward cyclic)
m2 <- tenmat(t, rdims = 2, cdims = "fc")
print(m2)

}
