% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_class.R
\name{tensor}
\alias{tensor}
\title{Create a tensor object}
\usage{
tensor(data, dims = NULL)
}
\arguments{
\item{data}{A vector, matrix, array, or numeric value}

\item{dims}{The dimensions of the tensor. If NULL, inferred from data}
}
\value{
A new Tensor object
}
\description{
Convenience function to create a Tensor object
}
\examples{
# Create a tensor from an array
t1 <- tensor(array(1:24, dim = c(3, 4, 2)))

# Create a tensor by specifying dimensions
t2 <- tensor(1:24, dims = c(3, 4, 2))

# Create a tensor filled with a single value
t3 <- tensor(0, dims = c(3, 4, 2))

}
