% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_class.R
\name{tensor}
\alias{tensor}
\title{Create a tensor object}
\usage{
tensor(data, dims = NULL)
}
\arguments{
\item{data}{A vector, matrix, array, or numeric value}

\item{dims}{The dimensions of the tensor. If NULL, inferred from data}
}
\value{
A new Tensor object
}
\description{
Convenience function to create a Tensor object
}
