% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_class.R
\name{tensor}
\alias{tensor}
\title{Create a `tensor` object}
\usage{
tensor(data, dims = NULL)
}
\arguments{
\item{data}{A vector, matrix or array.}

\item{dims}{The dimension of the tensor. When `dims` is not supplied, use the dimension of the data.}
}
\value{
A \code{tensor} object.
}
\description{
Create a `tensor` object
}
\examples{
tensor(data=matrix(1:6,ncol=2),dims=c(2,3))
tensor(data=matrix(c(1L,2L,3L,4L),ncol=2) )
mytensor <- tensor(data=array(1:1000000,c(100,100,100)))
mytensor + mytensor
}
