% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensory-package.R
\docType{package}
\name{tensory-package}
\alias{tensory}
\alias{tensory-package}
\title{tensory: Tenosry}
\description{
`tensory` is an expermental R package to implement tensor operations and functions based on `xtensor`.
}
\keyword{internal}
