% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttensor_class.R
\name{ttensor}
\alias{ttensor}
\title{Create a Tucker tensor (TTensor)}
\usage{
ttensor(core = NULL, U = NULL)
}
\arguments{
\item{core}{A Tensor object representing the core}

\item{U}{A list of factor matrices}
}
\value{
A TTensor object
}
\description{
Create a Tucker tensor (TTensor)
}
