% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_operations.R
\name{ttm}
\alias{ttm}
\title{Tensor Times Matrix (ttm) Operation}
\usage{
ttm(tensor, matrix, mode = 1, transpose = FALSE)
}
\arguments{
\item{tensor}{A Tensor object}

\item{matrix}{A matrix or list of matrices to multiply with the tensor}

\item{mode}{Integer specifying which mode (dimension) to multiply. 
For single matrix, defaults to 1. For multiple matrices, 
this should be a vector of modes corresponding to each matrix.}

\item{transpose}{Logical. If TRUE, transpose the matrix before multiplication.
Default is FALSE.}
}
\value{
A new Tensor object with the result of the multiplication
}
\description{
Compute a tensor times a matrix (or matrices) in one (or more) modes.
This function implements the tensor times matrix operation similar to MATLAB's ttm function.
Uses efficient Xtensor-blas implementation for high-performance tensor operations.
}
\details{
The tensor times matrix (ttm) operation multiplies a tensor by a matrix along a specified mode.
For a tensor X of size I1 x I2 x ... x IN and matrix M of size J x Ik, the result is a tensor
of size I1 x ... x I(k-1) x J x I(k+1) x ... x IN.
}
\examples{
# Create a 3D tensor
t <- tensor(array(1:24, dim = c(4, 3, 2)))

# Create a matrix
m <- matrix(1:8, nrow = 4, ncol = 2)

# Multiply tensor by matrix in mode 1
result <- ttm(t, m, mode = 1)

# Multiple matrices
m1 <- matrix(1:8, nrow = 4, ncol = 2)
m2 <- matrix(1:6, nrow = 3, ncol = 2)
result <- ttm(t, list(m1, m2), mode = c(1, 2))

}
