% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_operations.R
\name{ttt}
\alias{ttt}
\title{Tensor Times Tensor (ttt) Operation}
\usage{
ttt(tensorA, tensorB, dimsA = NULL, dimsB = dimsA)
}
\arguments{
\item{tensorA}{A Tensor object}

\item{tensorB}{A Tensor object}

\item{dimsA}{Subscripts of the dimensions in \code{tensorA} to contract over.}

\item{dimsB}{Subscripts of the dimensions in \code{tensorB} to contract over. Defaults to \code{dimsA}.}
}
\value{
A new Tensor object representing the product.
}
\description{
Computes the generalized product (outer, inner, or contracted) of two tensors.
}
\details{
\itemize{
\item If \code{dimsA} and \code{dimsB} are NULL, computes the outer product.
\item If \code{dimsA} and \code{dimsB} are provided, computes the contracted product along the specified dimensions.
The sizes of the corresponding dimensions must match.
}
}
\examples{
t1 <- tensor(array(1:24, dim = c(4, 3, 2)))
t2 <- tensor(array(1:12, dim = c(3, 2, 2)))

# Outer product
result_outer <- ttt(t1, t2)

# Contract over dimensions
result_contracted <- ttt(t1, t2, dimsA = c(2, 3), dimsB = c(1, 2))

}
